[Refactor] : 코드 스타일 및 디렉토리 구조 변경(#27)
